Comum/ContaCapitalComumEJB/src/test/java/br/com/sicoob/cca/comum/negocio/servicos/interceptors/ContaCapitalComumInternacionalizacaoInterceptorTest.java
package br.com.sicoob.cca.comum.negocio.servicos.interceptors;

import org.junit.Assert;
import org.junit.BeforeClass;
import org.junit.Test;

import br.com.bancoob.negocio.contexto.InformacoesUsuario;

public class ContaCapitalComumInternacionalizacaoInterceptorTest {

	@BeforeClass
	public static void setUpClass() {
		InformacoesUsuario.INSTANCIA.set(new InformacoesUsuario());
	}
	
	@Test
	public void deveInstanciarInterceptor() {
		ContaCapitalComumInternacionalizacaoInterceptor interceptor = 
				new ContaCapitalComumInternacionalizacaoInterceptor();
		Assert.assertNotNull(interceptor);
		Assert.assertNotNull(interceptor.getResourceBundle());
	}
	
}
